Update Storage-Spaces-Direct-Hardware-Requirements.md
Requirement for SES is ambiguous. I have seen multiple customer buy hardware without SES because the requirement was not clearly called out. Most people read the table of requirements and not the detail. There is no harm in calling this out in the bullet points to make sure a it is not missed and it could prevent customers from making a very expensive hardware mistake.